# LogicArray changes