--- a/docs/source/newsfragments/4093.removal.1.rst
+++ b/docs/source/newsfragments/4093.removal.1.rst
@@ -1,0 +1,1 @@
+Constructing a :class:`~cocotb.types.LogicArray` from an :class:`int` now only accepts integer literals (unsigned). Use :meth:`.LogicArray.from_signed` to convert negative integers into :class:`~cocotb.types.LogicArray`\ s using two's complement representation.