--- a/docs/source/newsfragments/4093.removal.rst
+++ b/docs/source/newsfragments/4093.removal.rst
@@ -1,0 +1,1 @@
+When constructing a :class:`~cocotb.types.LogicArray` from a :class:`int`, the *range* argument is now required.