--- a/examples/simple_dff/test_dff.py
+++ b/examples/simple_dff/test_dff.py
@@ -22,7 +22,7 @@ async def dff_simple_test(dut):
     # verilator does not support 4-state signals
     # see https://veripool.org/guide/latest/languages.html#unknown-states
     initial = (
-        LogicArray(0)
+        LogicArray("0")
         if cocotb.SIM_NAME.lower().startswith("verilator")
         else (
             LogicArray("U") if LANGUAGE.lower().startswith("vhdl") else LogicArray("X")