--- a/src/cocotb/types/array.py
+++ b/src/cocotb/types/array.py
@@ -1,12 +1,12 @@
 # Copyright cocotb contributors
 # Licensed under the Revised BSD License, see LICENSE for details.
 # SPDX-License-Identifier: BSD-3-Clause
-import typing
+from typing import Iterable, Iterator, Optional, TypeVar, Union, cast, overload
 
 from cocotb.types import ArrayLike
 from cocotb.types.range import Range
 
-T = typing.TypeVar("T")
+T = TypeVar("T")
 
 
 class Array(ArrayLike[T]):
@@ -127,7 +127,7 @@ class Array(ArrayLike[T]):
         "_range",
     )
 
-    def __init__(self, value: typing.Iterable[T], range: typing.Optional[Range] = None):
+    def __init__(self, value: Iterable[T], range: Optional[Range] = None):
         self._value = list(value)
         if range is None:
             self._range = Range(0, "to", len(self._value) - 1)
@@ -154,10 +154,10 @@ class Array(ArrayLike[T]):
             )
         self._range = new_range
 
-    def __iter__(self) -> typing.Iterator[T]:
+    def __iter__(self) -> Iterator[T]:
         return iter(self._value)
 
-    def __reversed__(self) -> typing.Iterator[T]:
+    def __reversed__(self) -> Iterator[T]:
         return reversed(self._value)
 
     def __contains__(self, item: object) -> bool:
@@ -171,15 +171,13 @@ class Array(ArrayLike[T]):
         else:
             return NotImplemented
 
-    @typing.overload
+    @overload
     def __getitem__(self, item: int) -> T: ...
 
-    @typing.overload
+    @overload
     def __getitem__(self, item: slice) -> "Array[T]": ...
 
-    def __getitem__(
-        self, item: typing.Union[int, slice]
-    ) -> typing.Union[T, "Array[T]"]:
+    def __getitem__(self, item: Union[int, slice]) -> Union[T, "Array[T]"]:
         if isinstance(item, int):
             idx = self._translate_index(item)
             return self._value[idx]
@@ -199,18 +197,18 @@ class Array(ArrayLike[T]):
             return Array(value=value, range=range)
         raise TypeError(f"indexes must be ints or slices, not {type(item).__name__}")
 
-    @typing.overload
+    @overload
     def __setitem__(self, item: int, value: T) -> None: ...
 
-    @typing.overload
-    def __setitem__(self, item: slice, value: typing.Iterable[T]) -> None: ...
+    @overload
+    def __setitem__(self, item: slice, value: Iterable[T]) -> None: ...
 
     def __setitem__(
-        self, item: typing.Union[int, slice], value: typing.Union[T, typing.Iterable[T]]
+        self, item: Union[int, slice], value: Union[T, Iterable[T]]
     ) -> None:
         if isinstance(item, int):
             idx = self._translate_index(item)
-            self._value[idx] = typing.cast(T, value)
+            self._value[idx] = cast(T, value)
         elif isinstance(item, slice):
             start = item.start if item.start is not None else self.left
             stop = item.stop if item.stop is not None else self.right
@@ -222,7 +220,7 @@ class Array(ArrayLike[T]):
                 raise IndexError(
                     f"slice [{start}:{stop}] direction does not match array direction [{self.left}:{self.right}]"
                 )
-            value = list(typing.cast(typing.Iterable[T], value))
+            value = list(cast(Iterable[T], value))
             if len(value) != (stop_i - start_i + 1):
                 raise ValueError(
                     f"value of length {len(value)!r} will not fit in slice [{start}:{stop}]"