--- a/src/cocotb/types/logic.py
+++ b/src/cocotb/types/logic.py
@@ -109,8 +109,6 @@ class Logic:
         ValueError: if the value cannot be constructed into a :class:`Logic`.
     """
 
-    __slots__ = ("_repr",)
-
     _repr: int
 
     @classmethod