--- a/tests/pytest/test_logic_array.py
+++ b/tests/pytest/test_logic_array.py
@@ -27,12 +27,8 @@ def test_logic_array_constructor():
 
 def test_logic_array_constructor_deprecated():
     with pytest.warns(DeprecationWarning):
-        assert LogicArray(0xA7) == LogicArray("10100111")
-    with pytest.warns(DeprecationWarning):
         assert LogicArray(10, Range(5, "downto", 0)) == LogicArray("001010")
 
-    with pytest.warns(DeprecationWarning):
-        assert LogicArray(-1) == LogicArray("1")
     with pytest.warns(DeprecationWarning):
         assert LogicArray(-2, Range(5, "downto", 0)) == LogicArray("111110")
 
@@ -45,15 +41,13 @@ def test_logic_array_constructor_deprecated():
 
 
 def test_logic_array_int_conversion():
-    assert LogicArray.from_unsigned(0xA7) == LogicArray("10100111")
     assert LogicArray.from_unsigned(10, Range(5, "downto", 0)) == LogicArray("001010")
-    with pytest.raises(OverflowError):
-        LogicArray.from_unsigned(-10)
+
     with pytest.raises(OverflowError):
         LogicArray.from_unsigned(10, Range(1, "to", 3))
 
-    assert LogicArray.from_signed(-1) == LogicArray("1")
     assert LogicArray.from_signed(-2, Range(5, "downto", 0)) == LogicArray("111110")
+
     with pytest.raises(OverflowError):
         LogicArray.from_signed(-45, Range(1, "to", 3))