… tests/test_cases/test_vhdl_zerovector/test_vhdl_zerovector.py