Removes some features that were of little value, required guessing, or that were deprecated (those can go in 1.9). The idea is that `LogicArray` has 3 different "literal" syntaxes: integer literals (`0b1010101`), string literals (`"01011010"`), and array/tuple literals (`[1, 0, 1, 0]`). These can be used in construction and comparison. Perhaps they will make it to the other operations someday.

Improves performance in the most common operations: construction from binstr and comparison to integer. Construction time is halfed and comparison is 10x'd.

```
#Before
> python -m timeit -s "from cocotb.types import LogicArray, Range" "LogicArray('10101010')"
100000 loops, best of 5: 2.82 usec per loop

> python -m timeit -s "from cocotb.types import LogicArray, Range" -s "a = LogicArray('10101010')" "a == 123"
200000 loops, best of 5: 1.54 usec per loop

# After
> python -m timeit -s "from cocotb.types import LogicArray, Range" "LogicArray('10101010')"
200000 loops, best of 5: 1.19 usec per loop

> python -m timeit -s "from cocotb.types import LogicArray, Range" -s "a = LogicArray('10101010')" "a == 123"
2000000 loops, best of 5: 171 nsec per loop
```