@ktbarrett 2024-08-13
Newest commit reduces construction time even more. For some reason, specifying the direction in the Range construction takes a couple hundred nanos.
```
> python -m timeit -s "from cocotb.types import LogicArray, Range" "LogicArray('10101010')"
200000 loops, best of 5: 1.1 usec per loop
```