@ktbarrett 2024-08-13
And the newest one uses a faster algo for checking only valid binstr characters are accepted.
```
> python -m timeit -s "from cocotb.types import LogicArray, Range" "LogicArray('10101010')"
500000 loops, best of 5: 794 nsec per loop
```