@ktbarrett 2024-08-13
Had to undo the change to remove specifying the direction in the constructor, when emptying strings are passed, null Ranges are necessary, and null Ranges can only be specified by giving a direction.