@ktbarrett 2024-08-13
Newest numbers on `LogicArray` creation from binstr
```
> python -m timeit -r 10 -s "from cocotb.types import LogicArray, Range" "LogicArray('10101010')"
500000 loops, best of 10: 867 nsec per loop
```