@ktbarrett 2024-08-14
Out implementaton of `cached_property` in `cocotb._py_compat` is straight up ***bad***.