@ktbarrett 2024-08-15
Scrub documentation of LogicArray, apparently we aren't testing the code in them anymore and some is out of date.